feat: Update package dependencies
This commit updates the package dependencies in the pi directory. It adds the xpress-rate-limit package with version ^7.3.0 as a new dependency. It also updates the @types/body-parser package to version 1.19.5 and the @types/cors package to version 2.8.17. These updates ensure that the project is using the latest stable versions of these packages.

Signed-off-by: David <davidfco.pozo@gmail.com>